feat!: tighten bot, client, and Telegram type validation
Validate client and bot configuration eagerly, extract framework-neutral
webhook HTTP handling, classify internal runtime failures separately,
and model
input profile photos as typed variants.

BREAKING CHANGE: Some builder/configuration setters now return Result,
webhook
helper APIs are stricter, and InputProfilePhoto construction/validation
is
typed rather than a generic JSON wrapper.

+++ crates/tele/src/client/layers/bootstrap.rs
@@ -381,7 +381,7 @@ where
                         policy.base_backoff,
                         policy.max_backoff,
                         attempt,
-                        policy.jitter_ratio,
+                        f64::from(policy.jitter_ratio),
                     )
                 });
                 tokio::time::sleep(delay).await;
@@ -430,7 +430,7 @@ where
                         policy.base_backoff,
                         policy.max_backoff,
                         attempt,
-                        policy.jitter_ratio,
+                        f64::from(policy.jitter_ratio),
                     )
                 });
                 std::thread::sleep(delay);
